refactor: shared authenticator class with the IAM related functions
